todo-api/phase2/requirements.txt
attrs==26.1.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
dnspython==2.8.0
flasgger==0.9.7.1
Flask==3.1.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mistune==3.2.0
packaging==26.0
pymongo==4.16.0
pyperclip==1.11.0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
urllib3==2.6.3
Werkzeug==3.1.8
